Move ILTransform out of src/tests, address Roslyn analyzer warnings
